refactor(frame): modularize handler pipeline
