+++ backend/src/agent-skills/structure-type/frame/__tests__/handler.test.mjs
@@ -1,0 +1,79 @@
+import { describe, expect, test } from '@jest/globals';
+import { handler } from '../../../../../dist/agent-skills/structure-type/frame/handler.js';
+import { detectFrameStructuralType } from '../../../../../dist/agent-skills/structure-type/frame/detect.js';
+import { mergeFrameState } from '../../../../../dist/agent-skills/structure-type/frame/merge.js';
+import { buildFrameQuestions } from '../../../../../dist/agent-skills/structure-type/frame/interaction.js';
+
+describe('frame handler composed modules', () => {
+  test('keeps sticky frame detection for follow-up messages', () => {
+    const match = detectFrameStructuralType({
+      message: '层高3.6m',
+      locale: 'zh',
+      currentState: {
+        inferredType: 'frame',
+        structuralTypeKey: 'frame',
+        supportLevel: 'supported',
+        updatedAt: 0,
+      },
+    });
+
+    expect(match?.skillId).toBe('frame');
+    expect(match?.mappedType).toBe('frame');
+  });
+
+  test('does not treat material and sections as critical blockers', () => {
+    const missing = handler.computeMissing({
+      inferredType: 'frame',
+      frameDimension: '2d',
+      storyCount: 2,
+      bayCount: 2,
+      storyHeightsM: [3, 3],
+      bayWidthsM: [6, 6],
+      floorLoads: [
+        { story: 1, verticalKN: 120, lateralXKN: 30 },
+        { story: 2, verticalKN: 120, lateralXKN: 30 },
+      ],
+      updatedAt: 0,
+    }, 'execution');
+
+    expect(missing.critical).toEqual([]);
+  });
+
+  test('keeps total-load wording in interaction questions', () => {
+    const [question] = buildFrameQuestions(
+      ['floorLoads'],
+      ['floorLoads'],
+      { inferredType: 'frame', frameDimension: '2d', updatedAt: 0 },
+      'zh',
+    );
+
+    expect(question.question).toContain('各层总荷载');
+    expect(question.question).not.toContain('节点荷载');
+  });
+
+  test('merges y-direction follow-up loads into existing 3d frame state', () => {
+    const state = mergeFrameState(
+      {
+        inferredType: 'frame',
+        frameDimension: '3d',
+        floorLoads: [
+          { story: 1, verticalKN: 90, lateralXKN: 18 },
+          { story: 2, verticalKN: 90, lateralXKN: 18 },
+        ],
+        updatedAt: 0,
+      },
+      {
+        inferredType: 'frame',
+        floorLoads: [
+          { story: 1, lateralYKN: 12 },
+          { story: 2, lateralYKN: 12 },
+        ],
+      },
+    );
+
+    expect(state.floorLoads).toEqual([
+      { story: 1, verticalKN: 90, lateralXKN: 18, lateralYKN: 12 },
+      { story: 2, verticalKN: 90, lateralXKN: 18, lateralYKN: 12 },
+    ]);
+  });
+});
