+++ backend/src/agent-skills/structure-type/frame/extract-natural.ts
@@ -1,57 +1,266 @@
 import { normalizeNumber, normalizePositiveInteger } from '../../../agent-runtime/fallback.js';
-import type { DraftExtraction, DraftState } from '../../../agent-runtime/types.js';
+import type { DraftExtraction, DraftFloorLoad, DraftState } from '../../../agent-runtime/types.js';
+import { normalizeSectionName, normalizeSteelGrade } from './model.js';
 
 const CHINESE_NUMERAL_MAP: Record<string, number> = {
-  '一': 1,
-  '二': 2,
-  '两': 2,
-  '三': 3,
-  '四': 4,
-  '五': 5,
-  '六': 6,
-  '七': 7,
-  '八': 8,
-  '九': 9,
-  '十': 10,
+  '一': 1, '二': 2, '两': 2, '三': 3, '四': 4,
+  '五': 5, '六': 6, '七': 7, '八': 8, '九': 9, '十': 10,
 };
 
 function parseLocalizedPositiveInt(raw: string | undefined): number | undefined {
   if (!raw) return undefined;
-  const direct = normalizePositiveInteger(raw);
+  const trimmed = raw.trim();
+  const direct = normalizePositiveInteger(trimmed);
   if (direct !== undefined) return direct;
-  if (raw === '十') return 10;
-  return CHINESE_NUMERAL_MAP[raw.trim()];
+  if (trimmed === '十') return 10;
+  if (trimmed.length === 2 && trimmed.startsWith('十')) {
+    const ones = CHINESE_NUMERAL_MAP[trimmed[1]];
+    return ones ? 10 + ones : undefined;
+  }
+  if (trimmed.length === 2 && trimmed.endsWith('十')) {
+    const tens = CHINESE_NUMERAL_MAP[trimmed[0]];
+    return tens ? tens * 10 : undefined;
+  }
+  return CHINESE_NUMERAL_MAP[trimmed];
+}
+
+function extractPositiveInt(text: string, patterns: RegExp[]): number | undefined {
+  for (const pattern of patterns) {
+    const match = text.match(pattern);
+    if (!match?.[1]) continue;
+    const value = parseLocalizedPositiveInt(match[1]);
+    if (value !== undefined) return value;
+  }
+  return undefined;
+}
+
+function extractScalar(text: string, patterns: RegExp[]): number | undefined {
+  for (const pattern of patterns) {
+    const match = text.match(pattern);
+    if (!match?.[1]) continue;
+    const value = normalizeNumber(match[1]);
+    if (value !== undefined && value > 0) return value;
+  }
+  return undefined;
+}
+
+function extractDirectionalLoadScalar(text: string, axis: 'x' | 'y'): number | undefined {
+  const axisToken = axis;
+  return extractScalar(text, [
+    new RegExp(`${axisToken}(?:方)?向(?:水平|横向|侧向)?(?:总)?荷载(?:都?是|均为|各为|分别为|分别取|取|按|为|是)?\\s*([0-9]+(?:\\.[0-9]+)?)\\s*(?:kn|千牛)`, 'i'),
+    new RegExp(`(?:水平|横向|侧向)?(?:总)?荷载(?:都?是|均为|各为|分别为|分别取|取|按|为|是)?[^\\n]{0,24}?${axisToken}(?:方)?向\\s*([0-9]+(?:\\.[0-9]+)?)\\s*(?:kn|千牛)`, 'i'),
+    new RegExp(`${axisToken}(?:方)?向\\s*([0-9]+(?:\\.[0-9]+)?)\\s*(?:kn|千牛)`, 'i'),
+  ]);
+}
+
+function shouldMirrorHorizontalLoadToBothAxes(
+  text: string,
+  existingState: DraftState | undefined,
+  inferred3d: boolean,
+): boolean {
+  if (!(inferred3d || existingState?.frameDimension === '3d')) return false;
+  return (
+    text.includes('水平方向荷载')
+    || text.includes('水平荷载都是')
+    || text.includes('水平荷载均为')
+    || text.includes('横向荷载两个方向')
+    || text.includes('侧向荷载两个方向')
+    || text.includes('两个方向都是')
+    || text.includes('horizontal loads')
+    || text.includes('水平总荷载')
+    || /x(?:、|\/|和|及)\s*y(?:方)?向.{0,5}各/.test(text)
+  );
 }
 
 function repeatScalar(count: number | undefined, value: number | undefined): number[] | undefined {
-  if (!count || value === undefined) return undefined;
+  if (!count || !value) return undefined;
   return Array.from({ length: count }, () => value);
 }
 
-function extractMatch(text: string, pattern: RegExp): string | undefined {
-  const match = text.match(pattern);
-  return match?.[1];
+function extractDirectionalSegment(text: string, axis: 'x' | 'y'): string {
+  const pattern = axis === 'x'
+    ? /x(?:方向|向)([\s\S]*?)(?=y(?:方向|向)|$)/i
+    : /y(?:方向|向)([\s\S]*?)$/i;
+  return text.match(pattern)?.[1] || '';
+}
+
+function extractSpanArray(segment: string): number[] | undefined {
+  if (!segment) return undefined;
+
+  const tryExtract = (pattern: RegExp): number[] | undefined => {
+    const match = segment.match(pattern);
+    if (!match?.[1]) return undefined;
+    const values = [...match[1].matchAll(/([\d.]+)(?=\s*(?:m|米))/gi)]
+      .map((capture) => Number.parseFloat(capture[1]))
+      .filter((value) => Number.isFinite(value) && value > 0 && value < 500);
+    return values.length >= 2 ? values : undefined;
+  };
+
+  const spanContextRe = /(?:跨度|各跨|bay\s*width)\s*(?:分别|各为|为|是)?\s*((?:[\d.]+\s*(?:m|米)\s*[、，,和\s]?\s*)+)/i;
+  const dividedRe = /(?:^|[，,、\s])分别\s*((?:[\d.]+\s*(?:m|米)\s*[、，,和\s]?\s*)+)/i;
+
+  return tryExtract(spanContextRe) ?? tryExtract(dividedRe);
+}
+
+function buildUniformFloorLoads(
+  storyCount: number | undefined,
+  verticalKN: number | undefined,
+  lateralXKN: number | undefined,
+  lateralYKN: number | undefined,
+): DraftFloorLoad[] | undefined {
+  if (!storyCount) return undefined;
+  if (verticalKN === undefined && lateralXKN === undefined && lateralYKN === undefined) return undefined;
+  return Array.from({ length: storyCount }, (_, index) => ({
+    story: index + 1,
+    verticalKN,
+    lateralXKN,
+    lateralYKN,
+  }));
+}
+
+function extractSteelGrade(text: string): string | undefined {
+  const withKeyword = text.match(
+    /(?:材料|钢材|钢种|牌号|采用|选用)[\s:：]*([Qq][0-9]{3,4}|[Ss][0-9]{3}|[Aa]36)/i,
+  );
+  if (withKeyword?.[1]) return normalizeSteelGrade(withKeyword[1]);
+
+  const gradeMatch = text.match(/(?:^|[^a-zA-Z0-9])([Qq](?:235|345|355|390|420))(?![0-9])/);
+  if (gradeMatch?.[1]) return normalizeSteelGrade(gradeMatch[1]);
+
+  const intlMatch = text.match(/(?:steel\s*grade|grade|material)\s*([Ss](?:235|275|355)|[Aa]36)\b/i);
+  if (intlMatch?.[1]) return normalizeSteelGrade(intlMatch[1]);
+
+  return undefined;
+}
+
+function extractSectionDesignation(text: string, role: 'column' | 'beam'): string | undefined {
+  const roleZh = role === 'column' ? '柱' : '梁';
+  const roleEn = role === 'column' ? 'column' : 'beam';
+  const sectionPattern = '[Hh][WwNn][0-9]+(?:[xX×][0-9]+){1,3}';
+
+  const withRoleBefore = new RegExp(`${roleZh}(?:截面|断面|型号|规格)?[\\s:：]*(${sectionPattern})`, 'i');
+  const withRoleAfter = new RegExp(`(${sectionPattern})\\s*${roleZh}`, 'i');
+  const withEnBefore = new RegExp(`${roleEn}\\s*section\\s*(${sectionPattern})`, 'i');
+
+  const m1 = text.match(withRoleBefore);
+  if (m1?.[1]) return normalizeSectionName(m1[1]);
+  const m2 = text.match(withRoleAfter);
+  if (m2?.[1]) return normalizeSectionName(m2[1]);
+  const m3 = text.match(withEnBefore);
+  if (m3?.[1]) return normalizeSectionName(m3[1]);
+  return undefined;
 }
 
 export function normalizeFrameNaturalPatch(message: string, existingState: DraftState | undefined): DraftExtraction {
   const text = message.toLowerCase();
 
-  const storyCount = parseLocalizedPositiveInt(extractMatch(text, /([0-9]+|[一二两三四五六七八九十]+)\s*层/i));
-  const xCount = parseLocalizedPositiveInt(extractMatch(text, /x(?:方向|向).*?([0-9]+|[一二两三四五六七八九十]+)\s*跨/i));
-  const yCount = parseLocalizedPositiveInt(extractMatch(text, /y(?:方向|向).*?([0-9]+|[一二两三四五六七八九十]+)\s*跨/i));
+  const storyCount = extractPositiveInt(text, [
+    /([0-9]+|[一二两三四五六七八九十]+)\s*层/i,
+    /([0-9]+|[一二两三四五六七八九十]+)\s*stories?/i,
+  ]);
+  const genericBayCount = extractPositiveInt(text, [
+    /([0-9]+|[一二两三四五六七八九十]+)\s*跨/i,
+    /([0-9]+|[一二两三四五六七八九十]+)\s*bays?/i,
+  ]);
 
-  const storyHeight = normalizeNumber(extractMatch(text, /每层(?:层高)?(?:都?是|为)?\s*([0-9]+(?:\.[0-9]+)?)\s*(?:m|米)/i));
-  const xSpan = normalizeNumber(extractMatch(text, /x(?:方向|向).*?(?:间隔|跨度|每跨)(?:也?是|都?是|为)?\s*([0-9]+(?:\.[0-9]+)?)\s*(?:m|米)/i));
-  const ySpan = normalizeNumber(extractMatch(text, /y(?:方向|向).*?(?:间隔|跨度|每跨)(?:也?是|都?是|为)?\s*([0-9]+(?:\.[0-9]+)?)\s*(?:m|米)/i));
+  const xSegment = extractDirectionalSegment(text, 'x');
+  const ySegment = extractDirectionalSegment(text, 'y');
+
+  const bayCountX = extractPositiveInt(xSegment, [
+    /([0-9]+|[一二两三四五六七八九十]+)\s*跨/i,
+    /([0-9]+|[一二两三四五六七八九十]+)\s*bays?/i,
+  ]);
+  const bayCountY = extractPositiveInt(ySegment, [
+    /([0-9]+|[一二两三四五六七八九十]+)\s*跨/i,
+    /([0-9]+|[一二两三四五六七八九十]+)\s*bays?/i,
+  ]);
+
+  const storyHeightScalar = extractScalar(text, [
+    /每层(?:层高)?(?:都?是|统一为|为|高)?\s*([0-9]+(?:\.[0-9]+)?)\s*(?:m|米)/i,
+    /层高(?:都?是|统一为|为)?\s*([0-9]+(?:\.[0-9]+)?)\s*(?:m|米)/i,
+  ]);
+
+  const xSpanArray = extractSpanArray(xSegment);
+  const ySpanArray = extractSpanArray(ySegment);
+  const xBayScalar = xSpanArray
+    ? undefined
+    : extractScalar(xSegment, [
+      /(?:间隔|跨度|每跨)(?:也?是|都?是|为)?\s*([0-9]+(?:\.[0-9]+)?)\s*(?:m|米)/i,
+    ]);
+  const yBayScalar = ySpanArray
+    ? undefined
+    : extractScalar(ySegment, [
+      /(?:间隔|跨度|每跨)(?:也?是|都?是|为)?\s*([0-9]+(?:\.[0-9]+)?)\s*(?:m|米)/i,
+    ]);
+  const genericBayScalar = extractScalar(text, [
+    /每跨(?:都?是|为)?\s*([0-9]+(?:\.[0-9]+)?)\s*(?:m|米)/i,
+    /跨度(?:都?是|也是|为)?\s*([0-9]+(?:\.[0-9]+)?)\s*(?:m|米)/i,
+    /间隔(?:都?是|也是|为)?\s*([0-9]+(?:\.[0-9]+)?)\s*(?:m|米)/i,
+  ]);
+
+  const verticalLoadKN = extractScalar(text, [
+    /(?:每层|各层)(?:节点)?(?:竖向|垂直|竖直|总)?(?:方向)?荷载(?:都?是|均为|为|是)?\s*([0-9]+(?:\.[0-9]+)?)\s*(?:kn|千牛)/i,
+    /(?:每层|各层)(?:竖向|垂直|竖直)\s*([0-9]+(?:\.[0-9]+)?)\s*(?:kn|千牛)/i,
+    /(?:竖向|垂直|竖直)荷载[^0-9]{0,10}(?:每层|各层)[^0-9]{0,5}([0-9]+(?:\.[0-9]+)?)\s*(?:kn|千牛)/i,
+    /(?:竖向|垂直|竖直)(?:方向)?(?:都?是|为|是)?\s*([0-9]+(?:\.[0-9]+)?)\s*(?:kn|千牛)/i,
+  ]);
+
+  const dualLateralLoadKN = extractScalar(text, [
+    /x(?:、|\/|和|及)\s*y(?:方)?向(?:水平|横向|侧向)?(?:总)?荷载(?:都?是|均为|各为|为|是)?\s*([0-9]+(?:\.[0-9]+)?)\s*(?:kn|千牛)/i,
+    /水平(?:总)?荷载[^0-9]{0,24}?x(?:方)?向(?:和|\/|、|及)\s*y(?:方)?向(?:都?是|均为|各为|各|为|是)?\s*([0-9]+(?:\.[0-9]+)?)\s*(?:kn|千牛)/i,
+    /水平(?:总)?荷载x(?:、|\/|和|及)\s*y(?:方)?向(?:水平|横向|侧向)?(?:都?是|均为|各为|各|为|是)?\s*([0-9]+(?:\.[0-9]+)?)\s*(?:kn|千牛)/i,
+    /x(?:方)?向(?:和|\/|、|及)\s*y(?:方)?向(?:都?是|均为|各为|分别为|分别取|为|是)\s*([0-9]+(?:\.[0-9]+)?)\s*(?:kn|千牛)/i,
+  ]);
+  const extractedLateralXLoadKN = dualLateralLoadKN ?? extractScalar(text, [
+    /(?:横向|侧向|水平)(?:总)?(?:方向)?荷载(?:两个方向)?(?:都?是|均为|都为|为|是)?\s*([0-9]+(?:\.[0-9]+)?)\s*(?:kn|千牛)/i,
+    /水平(?:总)?方向荷载(?:都?是|均为|为|是)?\s*([0-9]+(?:\.[0-9]+)?)\s*(?:kn|千牛)/i,
+    /(?:横向|侧向|水平)(?:总)?荷载(?:都?是|均为|为|是)?\s*([0-9]+(?:\.[0-9]+)?)\s*(?:kn|千牛)/i,
+  ]) ?? extractDirectionalLoadScalar(text, 'x');
+  const extractedLateralYLoadKN = dualLateralLoadKN ?? extractDirectionalLoadScalar(text, 'y');
+
+  const resolvedStoryCount = storyCount ?? existingState?.storyCount ?? existingState?.storyHeightsM?.length;
+  const resolvedBayCountX = bayCountX ?? existingState?.bayCountX;
+  const resolvedBayCountY = bayCountY ?? existingState?.bayCountY;
+
+  const inferred3d = text.includes('y方向')
+    || text.includes('y向')
+    || bayCountY !== undefined
+    || yBayScalar !== undefined
+    || ySpanArray !== undefined
+    || extractedLateralYLoadKN !== undefined;
+  const resolvedFrameDimension = inferred3d
+    ? '3d'
+    : (existingState?.frameDimension ?? (bayCountX !== undefined ? '3d' : undefined));
+  const mirrorHorizontalLoad = shouldMirrorHorizontalLoadToBothAxes(text, existingState, inferred3d);
+  const lateralXLoadKN = extractedLateralXLoadKN;
+  const lateralYLoadKN = extractedLateralYLoadKN ?? (mirrorHorizontalLoad ? extractedLateralXLoadKN : undefined);
+
+  const frameMaterial = extractSteelGrade(message) ?? extractSteelGrade(text);
+  const frameColumnSection = extractSectionDesignation(message, 'column');
+  const frameBeamSection = extractSectionDesignation(message, 'beam');
 
   return {
     inferredType: 'frame',
-    frameDimension: yCount !== undefined || ySpan !== undefined ? '3d' : existingState?.frameDimension,
+    frameDimension: resolvedFrameDimension,
     storyCount,
-    storyHeightsM: repeatScalar(storyCount ?? existingState?.storyCount, storyHeight),
-    bayCountX: xCount,
-    bayCountY: yCount,
-    bayWidthsXM: repeatScalar(xCount, xSpan),
-    bayWidthsYM: repeatScalar(yCount, ySpan),
+    bayCount: resolvedFrameDimension !== '3d' ? genericBayCount : undefined,
+    bayCountX: xSpanArray ? xSpanArray.length : bayCountX,
+    bayCountY: ySpanArray ? ySpanArray.length : bayCountY,
+    storyHeightsM: repeatScalar(resolvedStoryCount, storyHeightScalar),
+    bayWidthsM: resolvedFrameDimension !== '3d'
+      ? repeatScalar(genericBayCount ?? existingState?.bayCount, genericBayScalar)
+      : undefined,
+    bayWidthsXM: xSpanArray
+      ?? repeatScalar(resolvedBayCountX, xBayScalar ?? (resolvedFrameDimension === '3d' ? genericBayScalar : undefined)),
+    bayWidthsYM: ySpanArray ?? repeatScalar(resolvedBayCountY, yBayScalar),
+    floorLoads: buildUniformFloorLoads(
+      resolvedStoryCount,
+      verticalLoadKN,
+      lateralXLoadKN,
+      resolvedFrameDimension === '3d' ? lateralYLoadKN : undefined,
+    ),
+    ...(frameMaterial !== undefined && { frameMaterial }),
+    ...(frameColumnSection !== undefined && { frameColumnSection }),
+    ...(frameBeamSection !== undefined && { frameBeamSection }),
   };
 }
